refactor: remove db worker, remove global export
- The worker used in the databaseBuilder has no real impact on
  performance and created many problems for tests
- The db was exported as a global object, which is not a good practice.
  Now the databaseBuilder returns a new DB.
- Add a first test for an sql query, made possible by these changes.
