+++ components/unit/UnitQuery.vue
@@ -25,6 +25,7 @@
             v-else-if="sql"
             v-bind="{
               sql,
+              db,
               parameterName: defaultViewElements.parameterName,
               parameterKey: defaultViewElements.parameterKey
             }"
@@ -68,6 +69,7 @@
 
 <script>
 import FileManager from '~/utils/file-manager'
+import { DB } from '@/utils/sql'
 
 export default {
   props: {
@@ -102,6 +104,10 @@ export default {
     postprocessors: {
       type: Object,
       default: () => {}
+    },
+    db: {
+      type: DB,
+      default: null
     }
   },
   data() {
