+++ manifests/experiences/twitter/__tests__/database.test.js
@@ -1,4 +1,4 @@
-import createItems from '../file-parser'
+import { createItems, databaseBuilder } from '../database'
 import {
   adImpressions,
   adEngagements,
@@ -9,7 +9,7 @@ import preprocessors from '~/manifests/preprocessors'
 import FileManager from '~/utils/file-manager'
 import { mockFile } from '~/utils/__mocks__/file-manager-mock'
 
-test('the twitter database builder parses the files correctly', async () => {
+async function getFileManager() {
   const fileManager = new FileManager(
     {
       'data/ad-impressions.js': preprocessors.twitter,
@@ -26,6 +26,12 @@ test('the twitter database builder parses the files correctly', async () => {
     JSON.stringify(adEngagements)
   )
   await fileManager.init([fileImpressions, fileEngagements], true)
+
+  return fileManager
+}
+
+test('the twitter file parser returns the correct items for the database', async () => {
+  const fileManager = await getFileManager()
   const impressionsFile = JSON.parse(
     await fileManager.getPreprocessedText('data/ad-impressions.js')
   )
@@ -36,4 +42,23 @@ test('the twitter database builder parses the files correctly', async () => {
   const { adsItems, targetingItems } = results
   expect(adsItems).toEqual(adsItemsExpected)
   expect(targetingItems).toEqual(targetingItemsExpected)
+})
+
+test('the advertisers-per-day query returns the correct items', async () => {
+  const fileManager = await getFileManager()
+  const db = await databaseBuilder(fileManager)
+  const sql = `
+    SELECT advertiserName, DATE(time) AS date, COUNT(*) AS count
+    FROM twitterAds
+    GROUP BY advertiserName, DATE(time);
+  `
+  const result = db.select(sql)
+  const expected = {
+    headers: ['advertiserName', 'date', 'count'],
+    items: [
+      { advertiserName: 'Illumeably', date: '2021-04-15', count: 1 },
+      { advertiserName: 'PwC Switzerland', date: '2021-04-15', count: 1 }
+    ]
+  }
+  expect(result).toEqual(expected)
 })
