+++ manifests/experiences/twitter/database.js
@@ -1,9 +1,8 @@
-// eslint-disable-next-line import/default
-import ItemsWorker from './database.worker.js'
-import db from '@/utils/sql'
-import { runWorker } from '@/utils/utils'
+import { JSONPath } from 'jsonpath-plus'
+import { DB } from '~/utils/sql'
 
-export default async function databaseBuilder(fileManager) {
+export async function databaseBuilder(fileManager) {
+  const db = new DB()
   await db.init()
 
   db.create('twitterAds', [
@@ -26,9 +25,65 @@ export default async function databaseBuilder(fileManager) {
   const engagementsFile = JSON.parse(
     await fileManager.getPreprocessedText('data/ad-engagements.js')
   )
-  const workerInput = { impressionsFile, engagementsFile }
-  const workerOutput = await runWorker(new ItemsWorker(), workerInput)
-  const { adsItems, targetingItems } = workerOutput
+  const { adsItems, targetingItems } = createItems({
+    impressionsFile,
+    engagementsFile
+  })
   db.insert('twitterAds', adsItems)
   db.insert('twitterCriteria', targetingItems)
+  return db
+}
+
+export function createItems({ impressionsFile, engagementsFile }) {
+  const impressions = JSONPath({
+    path: '$.*.ad.adsUserData.adImpressions.impressions[*]',
+    json: impressionsFile
+  })
+  const engagements = JSONPath({
+    path: '$.*.ad.adsUserData.adEngagements.engagements[*].impressionAttributes',
+    json: engagementsFile
+  })
+  const adsItems = []
+  const targetingItems = []
+  let targetingItemsId = 0
+  let adsId = 0
+  const files = [
+    { values: impressions, engagement: 0 },
+    { values: engagements, engagement: 1 }
+  ]
+  for (const file of files) {
+    file.values.forEach(v => {
+      const tweetId = v.promotedTweetInfo ? v.promotedTweetInfo.tweetId : null
+      const advertiserName = v.advertiserInfo
+        ? v.advertiserInfo.advertiserName
+        : null
+      adsItems.push({
+        id: adsId,
+        tweetId,
+        advertiserName,
+        time: v.impressionTime,
+        engagement: file.engagement
+      })
+      const criteria = JSONPath({
+        path: '$.matchedTargetingCriteria[*]',
+        json: v
+      })
+      criteria.forEach(criterion => {
+        let targetingType = null
+        let targetingValue = null
+        if (criterion) {
+          targetingType = criterion.targetingType || null
+          targetingValue = criterion.targetingValue || null
+        }
+        targetingItems.push({
+          id: targetingItemsId++,
+          adId: adsId,
+          targetingType,
+          targetingValue
+        })
+      })
+      adsId++
+    })
+  }
+  return { adsItems, targetingItems }
 }
