--- manifests/experiences/twitter/database.worker.js
@@ -1,3 +1,0 @@
-import createItems from './file-parser'
-
-onmessage = message => postMessage(createItems(message.data))
