+++ manifests/experiences/twitter/index.js
@@ -1,5 +1,5 @@
 import customPipelines from './pipeline'
-import databaseBuilder from './database'
+import { databaseBuilder } from './database'
 import postprocessors from './postprocessors'
 
 export default {
