+++ utils/sql.js
@@ -161,5 +161,3 @@ export class DB {
     }
   }
 }
-
-export default new DB()
